// src/apps/operations/ticket-relevage/ticket-relevage.model.ts
export interface TicketRelevageModel {
    id: string;
    exportateur_id: string;
    campagne_id: string;
    speculation_id: string;
    booking_id: string;
    conteneur_id: string;
    plomb_id: string;
    transitaire_id: string;
    code_pesee: string;
    booking_ticket: string;
    compagnie_maritime: string;
    first_weighing: number;
    date_first_weighing: Date;
    hour_first_weighing: string;
    second_weighing: number;
    date_second_weighing: Date;
    hour_second_weighing: string;
    poids_declare: number;
    poids_net: number;
    poids_vgm: number;
    nbr_emb: number;
    tare_emb: number;
    tare_pal: number;
    tare_cnt: number;
    tare_hab: number;
    tare_cnt2: number;
    metode_controle_vgm: string;
    provenance: string;
    destination: string;
    chauffeur: string;
    navire: string;
    port_embarquement: string;
    destination_navire: string;
    magazin: string;
    num_contrat: string;
    num_connaissement: string;
    num_lot: string;
    num_livraison: string;
    num_dossier: string;
    num_bord: string;
    tracteur: string;
    remorque: string;
    peseur: string;
    agent_cci: string;
    validity: boolean;
    created: Date;
    created_by: string;
    updated: Date;
    updated_by: string;
    file: any;
}